bdms/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
django==4.2.25
djangorestframework==3.15.2
importlib-metadata==8.5.0
Markdown==3.7
sqlparse==0.5.3
typing-extensions==4.13.2
zipp==3.20.2
